Change ext check IP file
